+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Integrate Hilt Dependency Injection
 - Quest | CQL runner for G6PD
 - Quest | Extraction of resources for Patient, Condition, Encounter, Observations
+- Quest | Added G6PD login configuration
 
 - EIR | Added workflow configurations
 - EIR | Add Adverse Events feature to log any reactions after the first dose of vaccine using Structure map
@@ -40,6 +41,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Engine | Integrate Hilt Dependency Injection
 - Engine | Fixed login authentication issue.
 - ANC | Implement vital signs observation extraction from questionnaires
+- Engine | Implement language switcher on profile page
+- Quest | Add Swahili translations for patient registration questionnaire and activate language switcher
+- ANC | Activate language switcher on profile page
 
 ### Fixed
 - ANC | Resolved a bug in the Patient details : App crash when computing BMI